Refactor SingleStation into container
* Refactor SingleStation into container
* Move presentation code into StationStatus
* See #32
